+++ CHANGELOG.md
@@ -26,6 +26,10 @@ have been noticeable effect in typical simulations. In particular the per-thread
 data was already getting freed when the whole process exited, so it would only
 affect a process that created and terminated many threads over its lifetime.
 
+* Fixed a potential race condition when exiting managed threads that did not
+have the `clear_child_tid` attribute set. This is unlikely to have affected most
+software running under Shadow, since most thread APIs use this attribute.
+
 Raw changes since v2.5.0:
 
 * [Merged PRs](https://github.com/shadow/shadow/pulls?q=is%3Apr+merged%3A%3E2023-03-23T18%3A20-0400)
